refactor(tpm): move seed resolution into a testable lib module
Move recovery-seed parsing, file/stdin reading, generation, and
resolution out of the binary-only commands/tpm.rs into
storage::tpm::portable::provision, leaving the CLI provision command as
a thin wrapper that handles the interactive prompt and displays the seed
and warnings. The moved logic gains direct unit coverage (generate_seed,
resolve_seed) alongside the relocated parse/file/stdin tests.
